requirements.txt
fastapi
pydantic
passlib
python-jose
motor
pydantic-settings
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0

